breaking: Remove BoemlyThemeProvider from ContextProvider
BREAKING CHANGE:
The ContextProvider does not contain a BoemlyThemeProvider any more.
Users are now required to set a BoemlyThemeProvider by themselves. This
allows users to customize the fonts and colors that wixll be used by this
librarys components. The setup/migration is documented in the README
and the Storybook documentation.